chore: cleans up versions test config
